Create validators.py file on Customer app
